# openapi-user-service.yaml
openapi: 3.0.3
info:
  title: User Service API - QuickSpeak
  description: |
    API para gestión de usuarios, autenticación e idiomas del sistema QuickSpeak.

    Incluye:
    - Autenticación (login, registro, OAuth)
    - Gestión de usuarios (CRUD completo)
    - Catálogo de idiomas
    - Gestión de idiomas de usuario
  version: 1.0.0
  contact:
    name: QuickSpeak Team
    email: support@quickspeak.com
servers:
  - url: https://user-service-quickspeak.azurewebsites.net
    description: Servidor de producción en Azure
  - url: http://localhost:8082
    description: Servidor de desarrollo local
...
paths:
...
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: Login de usuario
      description: Autenticar usuario con email y contraseña, retorna JWT token
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@example.com
              password: password123
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/register:
    post:
      tags:
        - Authentication
      summary: Registro de usuario
      description: Registrar nuevo usuario y generar JWT token automáticamente
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
            example:
              email: newuser@example.com
              password: password123
              firstName: John
              lastName: Doe
      responses:
        '201':
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Usuario ya existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/oauth:
    post:
      tags:
        - Authentication
      summary: Login/Registro OAuth
      description: Autenticar o registrar usuario vía OAuth (Google, etc.)
      operationId: oauthLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthLoginRequest'
            example:
              email: user@gmail.com
              firstName: John
              lastName: Doe
              provider: GOOGLE
              providerId: google-user-id-123
      responses:
        '200':
          description: Usuario existente autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '201':
          description: Nuevo usuario creado y autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== USERS ====================
...
components:
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Email del usuario
          example: user@example.com
        password:
          type: string
          format: password
          description: Contraseña del usuario
          minLength: 6
          example: password123
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT token de autenticación
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        userId:
          type: integer
          format: int64
          description: ID del usuario
          example: 1
        email:
          type: string
          format: email
          description: Email del usuario
          example: user@example.com
        firstName:
          type: string
          description: Nombre del usuario
          example: John
        lastName:
          type: string
          description: Apellido del usuario
          example: Doe
    OAuthLoginRequest:
      type: object
      required:
        - email
        - firstName
        - lastName
        - provider
        - providerId
      properties:
        email:
          type: string
          format: email
          description: Email del usuario desde el proveedor OAuth
          example: user@gmail.com
        firstName:
          type: string
          description: Nombre del usuario
          example: John
        lastName:
          type: string
          description: Apellido del usuario
          example: Doe
        provider:
          type: string
          description: Proveedor OAuth (GOOGLE, FACEBOOK, etc.)
          enum:
            - GOOGLE
            - FACEBOOK
            - APPLE
          example: GOOGLE
        providerId:
          type: string
          description: ID único del usuario en el proveedor OAuth
          example: google-user-id-123

    # User Schemas
    UserRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
          description: Email del usuario
          example: user@example.com
        password:
          type: string
          format: password
          description: Contraseña del usuario
          minLength: 6
          example: password123
        firstName:
          type: string
          description: Nombre del usuario
          example: John
        lastName:
          type: string
          description: Apellido del usuario
          example: Doe
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Timestamp del error
          example: '2025-01-15T10:30:00Z'
        status:
          type: integer
          description: Código de estado HTTP
          example: 400
        error:
          type: string
          description: Nombre del error
          example: Bad Request
        message:
          type: string
          description: Mensaje de error descriptivo
          example: Invalid email format
        path:
          type: string
          description: Path del endpoint que generó el error
          example: /api/v1/auth/login
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtenido del endpoint de login

# Aplicar seguridad JWT a endpoints que lo requieran
# (Descomenta las líneas siguientes si necesitas proteger endpoints)
# security:
#   - bearerAuth: []