# src/main/java/com/yourteacher/userservice/adapter/in/web/dto/LoginResponse.java
package com.yourteacher.userservice.adapter.in.web.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

/**
 * DTO para respuesta de login
 */
@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class LoginResponse {

    private String token;
    private Long userId;
    private String email;
    private String firstName;
    private String lastName;
}